integerをfloatに変更

+++ db/migrate/20201212045729_create_dead_lift_weight_records.rb
@@ -1,7 +1,7 @@
 class CreateDeadLiftWeightRecords < ActiveRecord::Migration[6.0]
   def change
     create_table :dead_lift_weight_records do |t|
-      t.integer      :dead_lift_weight
+      t.float      :dead_lift_weight
       t.date       :dead_lift_day
       t.references :user,               foreign_key: true
       t.timestamps
